requirements.txt
langchain
langchain-openai
langchain_teddynote
pypdf
langchain_experimental
langchain-core
langchain-community
langgraph
langchain-opentutorial
pdfplumber
faiss-cpu
python-dotenv